Move model stuff to a separate package under src/
